Move settings to its own component
